now successfully pulling data, and organised the Account class some more
