+++ src/main/java/Account.java
@@ -7,7 +7,7 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
 
-public class Account {
+public class Account implements java.io.Serializable {
 
     public String username;
     public String firstName;
@@ -69,42 +69,5 @@ public class Account {
     }
 
     public Account() {
-    }
-
-    public static void firebaseInit() {
-        firebase = new Firebase("https://bank-of-myelin.firebaseio.com/");
-
-
-        // Initialize Firebase
-        try {
-            // [START initialize]
-            FirebaseOptions options = new FirebaseOptions.Builder()
-                    .setServiceAccount(new FileInputStream("service-account.json"))
-                    .setDatabaseUrl("https://bank-of-myelin.firebaseio.com/")
-                    .build();
-            FirebaseApp.initializeApp(options);
-            // [END initialize]
-        } catch (FileNotFoundException e) {
-            System.out.println("ERROR: invalid service account credentials. See README.");
-            System.out.println(e.getMessage());
-
-            System.exit(1);
-        }
-
-        DatabaseReference ref = FirebaseDatabase
-                .getInstance()
-                .getReference("Users/Accounts");
-        ref.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                Object document = dataSnapshot.getValue();
-                System.out.println(document);
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
     }
 }
